--- src/test/java/JAWAHER/User_Story_14/Test_Case3.java
@@ -1,4 +1,0 @@
-package JAWAHER.User_Story_14;
-
-public class Test_Case3 {
-}
