+++ src/test/java/J_User_Story_13/Test_Case1.java
@@ -1,11 +1,10 @@
-package JAWAHER.User_Story_13;
+package J_User_Story_13;
 //US_0013 Displays teams in the Teams module
 //Description:
 //1-Teams Are Displayed and Clickable
 //2-Add New Team
-
-import JAWAHER.Base_;
 import org.openqa.selenium.By;
+import org.openqa.selenium.Keys;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.chrome.ChromeDriver;
@@ -19,7 +18,7 @@ import java.time.Duration;
 import java.util.List;
 
 @Test
-public class Test_Case1 extends Base_ {
+public class Test_Case1  {
 WebDriver driver;
 Wait<WebDriver> wait;
 
@@ -104,6 +103,49 @@ Wait<WebDriver> wait;
 
     }
 
+
+    protected void ClickTeamsModule(){
+
+        WebDriverWait wait = new WebDriverWait(driver,Duration.ofSeconds(30));
+        wait.until(ExpectedConditions.visibilityOfElementLocated(By.xpath("//li[@id='link7']"))).click();;
+    }
+    protected void Click_AddNewButton (){
+
+        String xpath="//button[contains(@class, 'btn btn-info float-end text-white') and @type='button']";
+        wait = new WebDriverWait(driver, Duration.ofSeconds(30));
+        wait.until(ExpectedConditions.visibilityOfElementLocated(By.xpath(xpath))).click();
+
+    }
+
+    protected void AddNewTeamMethod() {
+        String Name="Quality Assurance 5";
+        String ShortcutName="QA 5";
+
+        //Name
+        wait = new WebDriverWait(driver, Duration.ofSeconds(30));
+        var Field1 =wait.until(ExpectedConditions.visibilityOfElementLocated(By.id("name")));
+        //wait.pollingEvery(Duration.ofMillis(2000));
+        Field1.sendKeys(Name);
+
+        //shortName
+        wait = new WebDriverWait(driver, Duration.ofSeconds(10));
+        var Field2 =wait.until(ExpectedConditions.visibilityOfElementLocated(By.name("short_name")));
+        Field2.sendKeys(ShortcutName);
+
+        //Team Type
+        String depType_dropdown_path = "//div[@id='react-select-3-placeholder']";//Locate element by ID
+        // Find the dropdown element
+        wait = new WebDriverWait(driver, Duration.ofSeconds(30));
+        WebElement depType_dropdown =wait.until(ExpectedConditions.visibilityOfElementLocated(By.id(depType_dropdown_path)));
+        //wait.pollingEvery(Duration.ofMillis(1000));
+        depType_dropdown.
+                sendKeys("Team", Keys.ENTER);
+
+        //Click Save Button
+        var SaveButton = driver.findElement(By.xpath("//button[@class=\"btn btn-info text-white px-3\"]"));
+        wait = new WebDriverWait(driver, Duration.ofSeconds(5));
+        SaveButton.click();
+    }
 
 
 }
