+++ src/test/java/J_User_Story_13/Test_Case3.java
@@ -1,4 +1,4 @@
-package JAWAHER.User_Story_13;
+package J_User_Story_13;
 
 public class Test_Case3 {
