+++ src/test/java/J_User_Story_14/Test_Case2.java
@@ -1,4 +1,4 @@
-package JAWAHER.User_Story_14;
+package J_User_Story_14;
 
 public class Test_Case2 {
